python_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pypdf==3.17.1
python-docx==1.1.0
ollama==0.4.4
python-dotenv==1.0.0
scikit-learn==1.6.0
numpy<2.0.0